handle update artist

+++ src/modules/image/image.service.ts
@@ -12,28 +12,38 @@ export class ImagesService {
     private ArtistRepository: ArtistRepository,
   ) {}
 
-  async updateImagesArtis(data: any, file: any): Promise<string> {
+  async updateImagesArtis(data: any, files: any): Promise<string> {
     try {
-      const { parentId, id } = data;
-      // Truy vấn Images cũ để lấy đường dẫn ảnh cũ từ db
-      const oldArtist = await this.ArtistRepository.findById(parentId);
-      if (!oldArtist) {
-        throw new Error('Images Artist not found');
-      }
-      //Lặp qua và xoá ảnh kèm theo trc khi updated
-      for (let index = 0; index < oldArtist.images.length; index++) {
-        const image = oldArtist.images[index];
-        if (image._id.toString() === id) {
-          await this.cloudinaryService.deleteImage(image.url);
+      const { parentId, id, status } = data;
+      let newUrl = '';
+      if (status == 0) {
+        //console.log(data);
+        // Truy vấn Images cũ để lấy đường dẫn ảnh cũ từ db
+        const oldArtist = await this.ArtistRepository.findById(parentId);
+        if (!oldArtist) {
+          throw new Error('Images Artist not found');
         }
+        //Lặp qua và xoá ảnh kèm theo trc khi updated
+        for (let index = 0; index < oldArtist.images.length; index++) {
+          const image = oldArtist.images[index];
+          if (image._id.toString() === id) {
+            await this.cloudinaryService.deleteImage(image.url);
+          }
+        }
+        //Ubload lên cloud và trả về url mới
+        const uploadImages = files.map((file: any) => {
+          return this.cloudinaryService.uploadImage(file);
+        });
+        const uploadResult = await Promise.all(uploadImages);
+        newUrl = uploadResult[0]?.secure_url;
+      } else {
+        //Ubload lên cloud và trả về url mới
+        const uploadImages = files.map((file: any) => {
+          return this.cloudinaryService.uploadImage(file);
+        });
+        const uploadResult = await Promise.all(uploadImages);
+        newUrl = uploadResult[0]?.secure_url;
       }
-
-      //Ubload lên cloud và trả về url mới
-      const uploadImages = file.map((file: any) => {
-        return this.cloudinaryService.uploadImage(file);
-      });
-      const uploadResult = await Promise.all(uploadImages);
-      const newUrl = uploadResult[0]?.secure_url;
 
       return newUrl;
     } catch (error) {
